backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain==0.3.14
langchain-community==0.3.14
langchain-chroma==0.2.0
chromadb==0.5.23
pypdf==5.1.0
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy==2.0.36
python-multipart==0.0.19
httpx==0.28.1
aiofiles==24.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
